refactor(services): add abstraction layer with interface/impl separation
- Convert all services to abstract interface classes (DeviceService, ModbusService,
  ChartsSettingsService, MeasUnitService, ScanService)
- Move concrete implementations to *_impl.dart files
- Inject ModbusService into DeviceServiceImpl via constructor (Riverpod DI)
- Fix write* methods in DeviceServiceImpl: c.name == c.name → c.name == device.name

fix(android): add debug AndroidManifest with INTERNET permission for Dart VM service
fix(android): bump Kotlin version 1.8.22 → 2.1.0
fix(ble): request BLUETOOTH_SCAN and BLUETOOTH_CONNECT at runtime before scanning
fix(ui): prevent bottom overflow in IndicationItemWidget using AutoSizeText for value
fix(ui): reduce padding and simplify FittedBox layout in MeasResultWidget
chore: add permission_handler dependency
