+++ idensity_ble_client/.gitignore
@@ -1,3 +1,6 @@
+CLAUDE.md
+.vscode/
+
 # Miscellaneous
 *.class
 *.log
