+++ idensity_ble_client/android/app/src/debug/AndroidManifest.xml
@@ -1,0 +1,4 @@
+<manifest xmlns:android="http://schemas.android.com/apk/res/android">
+    <!-- Required for Dart VM service (debugger, hot reload) -->
+    <uses-permission android:name="android.permission.INTERNET"/>
+</manifest>
