+++ idensity_ble_client/lib/models/providers/services_registration.dart
@@ -11,54 +11,59 @@ import 'package:idensity_ble_client/models/meas_units/meas_unit.dart';
 import 'package:idensity_ble_client/models/settings/meas_process.dart';
 import 'package:idensity_ble_client/services/bluetooth/ble_scan_service.dart';
 import 'package:idensity_ble_client/services/charts/charts_settings_service.dart';
+import 'package:idensity_ble_client/services/charts/charts_settings_service_impl.dart';
 import 'package:idensity_ble_client/services/device_service.dart';
+import 'package:idensity_ble_client/services/device_service_impl.dart';
 import 'package:idensity_ble_client/services/ethernet/ethernet_scan_service.dart';
 import 'package:idensity_ble_client/services/meas_units/meas_unit_service.dart';
+import 'package:idensity_ble_client/services/meas_units/meas_unit_service_impl.dart';
 import 'package:idensity_ble_client/services/modbus/modbus_service.dart';
+import 'package:idensity_ble_client/services/modbus/modbus_service_impl.dart';
 import 'package:idensity_ble_client/services/scan_service.dart';
 
 final appBarActionsProvider = StateProvider.autoDispose<List<Widget>>(
   (ref) => [],
 );
 
-final scanServiceProvider = FutureProvider.family
-    .autoDispose<ScanService, ConnectionType>((ref, conType) async {
-      final deviceService = await ref.watch(deviceServiceProvider.future);
-      ScanService? service;
-      switch (conType) {
-        case ConnectionType.bluetooth:
-          service = BleScanService(deviceService: deviceService);
-        default:
-          service = EthernetScanService(deviceService: deviceService);
-      }
-      ref.onDispose(() {
-        service?.dispose();
-      });
-      return service;
-    });
+final modbusServiceProvider = Provider<ModbusService>((ref) {
+  return ModbusServiceImpl();
+});
 
-final connectionTypeProvider = StateProvider<ConnectionType>(
-  (ref) => ConnectionType.bluetooth,
-);
-
-final deviceServiceProvider = FutureProvider<DeviceService>((ref) async{
+final deviceServiceProvider = FutureProvider<DeviceService>((ref) async {
   final logsRepo = ref.read(dataLogCellsRepositoryProvider);
   final deviceRepo = ref.read(deviceRepositoryProvider);
-  final service = DeviceService(
+  final modbusService = ref.read(modbusServiceProvider);
+  final service = DeviceServiceImpl(
     logCellRepository: logsRepo,
     deviceRepository: deviceRepo,
+    modbusService: modbusService,
   );
   ref.onDispose(() => service.dispose());
   await service.init();
   return service;
 });
 
-final modbusServiceProvider = Provider<ModbusService>((ref) {
-  return ModbusService();
+final scanServiceProvider = FutureProvider.family
+    .autoDispose<ScanService, ConnectionType>((ref, conType) async {
+  final deviceService = await ref.watch(deviceServiceProvider.future);
+  ScanService? service;
+  switch (conType) {
+    case ConnectionType.bluetooth:
+      service = BleScanService(deviceService: deviceService);
+    default:
+      service = EthernetScanService(deviceService: deviceService);
+  }
+  ref.onDispose(() {
+    service?.dispose();
+  });
+  return service;
 });
 
+final connectionTypeProvider = StateProvider<ConnectionType>(
+  (ref) => ConnectionType.bluetooth,
+);
+
 final deviceUpdateProvider = StreamProvider<Device>((ref) {
-  // Получаем экземпляр сервиса.
   final serviceAsyncValue = ref.watch(deviceServiceProvider);
   if (serviceAsyncValue.hasValue) {
     final service = serviceAsyncValue.value!;
@@ -67,20 +72,16 @@ final deviceUpdateProvider = StreamProvider<Device>((ref) {
   if (serviceAsyncValue.isLoading) {
     return const Stream.empty();
   } else {
-    throw Exception('Error loading  device service');
+    throw Exception('Error loading device service');
   }
 });
 
 final measUnitServiceProvider = FutureProvider<MeasUnitService>((ref) async {
   final repo = ref.read(measUnitsRepositoryProvider);
-  final service = MeasUnitService(measUnitRepository: repo);
-
-  // Регистрация метода dispose() для автоматического вызова
-  // когда провайдер будет уничтожен
+  final service = MeasUnitServiceImpl(measUnitRepository: repo);
   ref.onDispose(() {
     service.dispose();
   });
-
   await service.init();
   return service;
 });
@@ -88,12 +89,8 @@ final measUnitServiceProvider = FutureProvider<MeasUnitService>((ref) async {
 final chartSettingsServiceProvider = FutureProvider<ChartsSettingsService>((
   ref,
 ) async {
-  final deviceService = await ref.watch(deviceServiceProvider.future);
-
   final repo = ref.read(chartSettingsRepositoryProvider);
-
-  final service = ChartsSettingsService(
-    deviceService: deviceService,
+  final service = ChartsSettingsServiceImpl(
     chartSettingsRepository: repo,
   );
   await service.init();
@@ -106,7 +103,6 @@ final chartSettingsStreamProvider = StreamProvider<List<ChartSettings>>((ref) {
     final service = serviceAsyncValue.value!;
     return service.settingsStream;
   }
-
   if (serviceAsyncValue.isLoading) {
     return const Stream.empty();
   } else {
@@ -116,7 +112,6 @@ final chartSettingsStreamProvider = StreamProvider<List<ChartSettings>>((ref) {
 
 final measUnitsStreamProvider = StreamProvider<List<MeasUnit>>((ref) {
   final serviceAsyncValue = ref.watch(measUnitServiceProvider);
-
   if (serviceAsyncValue.hasValue) {
     final service = serviceAsyncValue.value!;
     return service.measUnitsStream;
@@ -165,15 +160,10 @@ final selectedMeasProcProvider = Provider<MeasProcess?>((ref) {
 
 final changeMeasUnitSelectingProvider = StreamProvider<Map<String, int>>((ref) {
   final serviceAsyncValue = ref.watch(measUnitServiceProvider);
-
-  // Проверяем, что значение готово
   if (serviceAsyncValue.hasValue) {
     final service = serviceAsyncValue.value!;
     return service.measUnitSelectingStream;
   }
-
-  // Если еще не готово, то остаемся в состоянии загрузки
-  // (или ошибки)
   if (serviceAsyncValue.isLoading) {
     return const Stream.empty();
   } else {
