+++ idensity_ble_client/lib/services/modbus/modbus_service.dart
@@ -1,243 +1,14 @@
-import 'dart:math';
-import 'dart:typed_data';
-
 import 'package:idensity_ble_client/models/connection.dart';
-import 'package:idensity_ble_client/models/connection_type.dart';
 import 'package:idensity_ble_client/models/indication/indication.dart';
-import 'package:idensity_ble_client/models/modbus/modbus_commands.dart';
 import 'package:idensity_ble_client/models/settings/device_settings.dart';
-import 'package:idensity_ble_client/services/modbus/extensions/data_indication_extensions.dart';
-import 'package:idensity_ble_client/services/modbus/extensions/device_settings_extensions.dart';
 
-class ModbusService {
-  static const int maxRegisterSize = 100;
-  final List<int> inputBuffer = List.filled(1000, 0);
+abstract interface class ModbusService {
+  Future<IndicationData> getIndicationData(Connection connection);
+  Future<DeviceSettings> getDeviceSettings(Connection connection);
 
-  Future<void> writeDeviceType(int value, Connection connection) async {
-    await _writeHoldingRegisters(
-      connection: connection,
-      registers: [value],
-      startAddr: 102,
-      count: 1,
-    );
-  }
-
-
-  Future<void> writeCalcType(int value, int measProcIndex, Connection connection)async{
-    await _writeHoldingRegisters(
-      connection: connection,
-      registers: [value],
-      startAddr: 204 + measProcIndex*180,
-      count: 1,
-    );
-  }
-
-  Future<void> writeMeasType(int value, int measProcIndex, Connection connection)async{
-    await _writeHoldingRegisters(
-      connection: connection,
-      registers: [value],
-      startAddr: 205 + measProcIndex*180,
-      count: 1,
-    );
-  }
-
-  Future<void> writeAveragePoints(int value, int measProcIndex, Connection connection) async {
-    await _writeHoldingRegisters(
-      connection: connection,
-      registers: [value],
-      startAddr: 200 + measProcIndex*180 + 1,
-      count: 1,
-    );
-  }
-
-  Future<void> writeMeasDuration(double value, int measProcIndex, Connection connection)async{
-    await _writeHoldingRegisters(
-      connection: connection,
-      registers: [(value*10).toInt()],
-      startAddr: 200 + measProcIndex*180,
-      count: 1,
-    );
-  }
-
-  Future<IndicationData> getIndicationData(Connection connection) async {
-    if (connection.connectionSettings.connectionType ==
-        ConnectionType.bluetooth) {
-      await _readInputRegisters(
-        connection: connection,
-        startAddr: 0,
-        count: 60,
-      );
-      final IndicationData indicationData = IndicationData();
-      indicationData.updateDataFromModbus(inputBuffer);
-      return indicationData;
-    } else {
-      throw Exception(
-        "Modbus service: ethernet interface is not implemented yet",
-      );
-    }
-  }
-
-  Future<DeviceSettings> getDeviceSettings(Connection connection) async {
-    if (connection.connectionSettings.connectionType ==
-        ConnectionType.bluetooth) {
-      await _readHoldingRegisters(
-        connection: connection,
-        startAddr: 0,
-        count: 400,
-      );
-      final DeviceSettings deviceSettings = DeviceSettings();
-      deviceSettings.updateDataFromModbus(inputBuffer);
-      return deviceSettings;
-    } else {
-      throw Exception(
-        "Modbus service: ethernet interface is not implemented yet",
-      );
-    }
-  }
-
-  Future<void> _readInputRegisters({
-    required Connection connection,
-    required int startAddr,
-    required int count,
-    int unitId = 1,
-  }) async {
-    await _readRegsitersCommon(
-      connection: connection,
-      command: ModbusReadCommands.readInputRegisters,
-      startAddr: startAddr,
-      count: count,
-      unitId: unitId,
-    );
-  }
-
-  Future<void> _readHoldingRegisters({
-    required Connection connection,
-    required int startAddr,
-    required int count,
-    int unitId = 1,
-  }) async {
-    await _readRegsitersCommon(
-      connection: connection,
-      command: ModbusReadCommands.readHoldingRegisters,
-      startAddr: startAddr,
-      count: count,
-      unitId: unitId,
-    );
-  }
-
-  Future<void> _readRegisters({
-    required Connection connection,
-    required ModbusReadCommands command,
-    required int startAddr,
-    required int count,
-    int unitId = 1,
-  }) async {
-    final request = Uint8List(8);
-    request[0] = unitId;
-    request[1] = command.code;
-    final view = ByteData.view(request.buffer);
-    view.setUint16(2, startAddr, Endian.big);
-    view.setUint16(4, count, Endian.big);
-    final crc = calculateCrc16(request, 6);
-    view.setUint16(6, crc, Endian.little);
-
-    var responce = await connection.readBytes(request, expectedRespLen: 5+count*2);
-    if (responce.length != (count * 2 + 5)) {
-      throw Exception("Invalid response length");
-    }
-    if (responce[1] != command.code) {
-      throw Exception("invalid request");
-    }
-    final byteList = Uint8List.fromList(responce);
-    var responceCrc = calculateCrc16(byteList, responce.length - 2);
-    var realCrc =
-        (responce[responce.length - 1] << 8) | responce[responce.length - 2];
-    if (realCrc != responceCrc) {
-      throw Exception("Crc responce err");
-    }
-    final ByteData byteData = ByteData.sublistView(byteList);
-    final end = startAddr + count;
-    for (var i = startAddr; i < end; i++) {
-      inputBuffer[i] = byteData.getUint16((i - startAddr) * 2 + 3, Endian.big);
-    }
-  }
-
-  Future<void> _writeHoldingRegisters({
-    required Connection connection,
-    required List<int> registers,
-    required int startAddr,
-    required int count,
-    int unitId = 1,
-  }) async {
-    final request = Uint8List(9 + count * 2);
-    request[0] = unitId;
-    request[1] = 16;
-    final view = ByteData.view(request.buffer);
-    view.setUint16(2, startAddr, Endian.big);
-    view.setUint16(4, count, Endian.big);
-    request[6] = count * 2;
-    for (var i = 0; i < count; i++) {
-      view.setUint16(7 + i * 2, registers[i], Endian.big);
-    }
-    final crc = calculateCrc16(request, 7 + count * 2);
-    view.setUint16(7 + count * 2, crc, Endian.little);
-
-    var responce = await connection.readBytes(request, expectedRespLen: 5+count*2);
-    if (responce.length != 8) throw Exception("Invalid response length");
-    if (responce[1] != 16) {
-      throw Exception("Write error to write data for modbusId = $unitId");
-    }
-    final byteList = Uint8List.fromList(responce);
-    var responceCrc = calculateCrc16(byteList, responce.length - 2);
-    var realCrc =
-        (responce[responce.length - 1] << 8) | responce[responce.length - 2];
-    if (realCrc != responceCrc) {
-      throw Exception("Crc responce err");
-    }
-  }
-
-  int calculateCrc16(Uint8List data, int length) {
-    int crc = 0xFFFF;
-    final polynomial = 0xA001; // Обратный полином 0x8005
-
-    for (var i = 0; i < length; i++) {
-      crc ^= data[i];
-      for (int i = 0; i < 8; i++) {
-        if ((crc & 0x0001) != 0) {
-          crc >>= 1;
-          crc ^= polynomial;
-        } else {
-          crc >>= 1;
-        }
-      }
-    }
-
-    return crc;
-  }
-
-  Future<void> _readRegsitersCommon({
-    required Connection connection,
-    required ModbusReadCommands command,
-    required int startAddr,
-    required int count,
-    int unitId = 1,
-  }) async {
-    final steps =
-        (count % maxRegisterSize == 0
-                ? count / maxRegisterSize
-                : count / maxRegisterSize + 1)
-            .toInt();
-    int start = startAddr;
-    for (var i = 0; i < steps; i++) {
-      int tmpCnt = min(maxRegisterSize, count - (i * maxRegisterSize));
-      await _readRegisters(
-        connection: connection,
-        command: command,
-        startAddr: start,
-        count: tmpCnt,
-        unitId: unitId,
-      );
-      start += tmpCnt;
-    }
-  }
+  Future<void> writeDeviceType(int value, Connection connection);
+  Future<void> writeMeasDuration(double value, int measProcIndex, Connection connection);
+  Future<void> writeAveragePoints(int value, int measProcIndex, Connection connection);
+  Future<void> writeCalcType(int value, int measProcIndex, Connection connection);
+  Future<void> writeMeasType(int value, int measProcIndex, Connection connection);
 }
