+++ idensity_ble_client/lib/services/scan_service.dart
@@ -1,24 +1,12 @@
 import 'package:idensity_ble_client/models/scan_result.dart';
 import 'package:idensity_ble_client/models/scan_state.dart';
 
-class ScanService {
-  List<IdensityScanResult> get scanResults => throw Exception("Needs to be implemented");
+abstract interface class ScanService {
+  List<IdensityScanResult> get scanResults;
+  Stream<ScanState> get scanState;
 
-  Future<void> stopScan() async {
-    throw Exception("Needs to be implemented");
-  }
-
-  Future<void> startScan({required int duration}) async {
-    throw Exception("Needs to be implemented");
-  }
-
-  Stream<ScanState> get scanState => throw Exception("Needs to be implemented");
-
-  void dispose() {
-    throw Exception("Needs to be implemented");
-  }
-
-  Future<void> saveDevices(List<IdensityScanResult> results) {
-    throw Exception("Needs to be implemented");
-  }
+  Future<void> startScan({required int duration});
+  Future<void> stopScan();
+  Future<void> saveDevices(List<IdensityScanResult> results);
+  void dispose();
 }
