+++ idensity_ble_client/pubspec.yaml
@@ -47,6 +47,7 @@ dependencies:
   flutter_colorpicker: ^1.1.0
   syncfusion_flutter_charts: ^32.1.19
   drift: ^2.30.0
+  permission_handler: ^11.4.0
   drift_flutter: ^0.2.8
