+++ idensity_ble_client/windows/flutter/generated_plugins.cmake
@@ -3,6 +3,7 @@
 #
 
 list(APPEND FLUTTER_PLUGIN_LIST
+  permission_handler_windows
   sqlite3_flutter_libs
   universal_ble
 )
